Convert relevant id references to long (#77)
* Fix header names for rate limiting

* Change Action ResourceId to long

* Convert all relevant id references to long

+++ DigitalOcean.API.Tests/Clients/ActionsClientTest.cs
@@ -15,7 +15,7 @@ namespace DigitalOcean.API.Tests.Clients {
 
             actionClient.Get(9001);
 
-            var parameters = Arg.Is<List<Parameter>>(list => (int)list[0].Value == 9001);
+            var parameters = Arg.Is<List<Parameter>>(list => (long)list[0].Value == 9001);
             factory.Received().ExecuteRequest<Action>("actions/{id}", parameters, null, "action");
         }
